Make changes to healthcheck
Add a flag to process.env to store the state of the database (whether
new blocks are being added or not). A message is sent to Slack whenever
the state changes. The check interval was increased to 70 seconds.
Slack channel and token are now stored in .env instead of config.